+++ Aurora/Source/Aurora/Utils/PlatformUtils.h
@@ -13,7 +13,7 @@ namespace Aurora {
 
 	class Time {
 	public:
-		static float GetTime();
+		static double GetTime();
 	};
 
 }
