+++ Aurora/Source/Platform/Windows/WindowsPlatformUtils.cpp
@@ -9,7 +9,7 @@
 
 namespace Aurora {
 
-	float Time::GetTime() {
+	double Time::GetTime() {
 		return glfwGetTime();
 	}
